+++ src/main/java/com/mamiyaotaru/voxelmap/util/BackgroundImageInfo.java
@@ -17,7 +17,7 @@ public class BackgroundImageInfo {
 
     public BackgroundImageInfo(BufferedImage image, int left, int top, int width, int height) {
         this.image = image;
-        this.glid = GLUtils.tex(image);
+        this.glid = OpenGL.Utils.tex(image);
         this.left = left;
         this.top = top;
         this.right = left + width;